all functionality added, some refactoring

+++ ShoppingList.js
@@ -6,82 +6,39 @@ const addButton = document.querySelector(".inputBox button");
 const shoppingList = document.querySelector(".shoppingItems");
 const deleteAll = document.querySelector(".footer button");
 const howManyPending = document.querySelector(".pendingTasks");
-//const bt = document.getElementById("submit");
-
-// const btn = document.getElementById("myBtn")
-// const input = document.getElementById("myInput")
-// inputBox.addEventListener("keyup", function(event) {
-//     if (event.key === 'Enter') {
-//         event.preventDefault();
-//         addButton.click();
-//     }
-//  });
-
-//||
 
 inputBox.onkeyup = () => {
   let elements = inputBox.value;
-  if (elements.trim() !== 0) {
-    addButton.classList.add("active");
-  }
-  inputBox.addEventListener("keydown", function (event) {
-    if (event.key === "Enter") {
-      //event.preventDefault();
-      addButton.click();
-    }
+  if (elements.trim() !== 0) addButton.classList.add("active");
+  inputBox.addEventListener("keydown", (event) => {
+    if (event.key === "Enter") addButton.click();
   });
 };
 
-// inputBox.onkeyup = () => {
-//     let elements = inputBox.value;
-//         if (elements.trim() !== 0) {
-//             addButton.classList.add("active");
-
-//     }
-// }
-
-//howManyPending.textContent = 0;
-whatIsThis();
+populateList();
 
 addButton.onclick = () => {
   let userInput = inputBox.value;
-  console.log(userInput);
   let lclStr = localStorage.getItem("new");
-  let myList = [];
-  console.log("clicked");
-  //if (lclStr) {
-  let test = JSON.parse(lclStr);
-  myList = [myList, ...test];
-  //}
-  console.log(myList);
-  // if (inputBox.value !== "") {
-  myList.push(userInput);
-  // }
+  let temp = JSON.parse(lclStr);
 
-  localStorage.setItem("new", JSON.stringify(myList));
-  console.log(localStorage);
-  console.log(myList);
-  whatIsThis();
+  if (inputBox.value !== "") temp.push(userInput);
+
+  localStorage.setItem("new", JSON.stringify(temp));
+  populateList();
 };
 
-function whatIsThis() {
-  let myList = [];
+function populateList() {
+  let temp = null;
   let getNewStuff = localStorage.getItem("new");
-  if (getNewStuff == null) {
-    myList = [];
-  } else {
-    myList = JSON.parse(getNewStuff);
-  }
+  getNewStuff == null ? (temp = "") : (temp = JSON.parse(getNewStuff));
 
-  howManyPending.textContent = myList.length;
+  howManyPending.textContent = temp.length;
 
-  if (myList.length > 0) {
-    deleteAll.classList.add("active");
-    // } else {
-    //     deleteButton.classList.remove("active");
-  }
+  if (temp.length > 0) deleteAll.classList.add("active");
+
   let listItemTag = "";
-  myList.forEach((element, index) => {
+  temp.forEach((element, index) => {
     listItemTag += `<li>${element}<span class="icon" onclick="deleteTask(${index})"><i class="fas fa-trash"></i></span></li>`;
   });
   shoppingList.innerHTML = listItemTag;
@@ -90,14 +47,14 @@ function whatIsThis() {
 
 let deleteTask = (param) => {
   let lclStr = localStorage.getItem("new");
-  let myList = JSON.parse(lclStr);
-  myList.splice(param, 1);
-  localStorage.setItem("new", JSON.stringify(myList));
-  whatIsThis();
+  let temp = JSON.parse(lclStr);
+  temp.splice(param, 1);
+  localStorage.setItem("new", JSON.stringify(temp));
+  populateList();
 };
 
 deleteAll.onclick = () => {
-  let myList = []; //empty the array
-  localStorage.setItem("new", JSON.stringify(myList)); //set the item in localstorage
-  whatIsThis(); //call the showTasks function
+  let temp = [];
+  localStorage.setItem("new", JSON.stringify(temp));
+  populateList();
 };
